+++ src/Client/ModuleNode.cpp
@@ -67,6 +67,16 @@ void ModuleNode::setup()
 
         this->addChildAtPosition(optionsBtn, Anchor::Left, ccp(label->getScaledContentWidth() + 43, 0));
     }
+
+    if (hasDesc)
+    {
+        // tooltips integration
+        // auto tooltipBtn = CCMenuItemSpriteExtra::create(CCSprite::createWithSpriteFrameName("GJ_infoIcon_001.png"), this, nullptr);
+        // tooltipBtn->setAnchorPoint(ccp(0, 0));
+        // tooltipBtn->setContentSize(getContentSize());
+        // this->addChild(tooltipBtn);
+        // tooltipBtn->setUserObject("alphalaneous.tooltips/tooltip", CCString::create("ASDFUISDFGSDFGIYU"));
+    }
 }
 
 void ModuleNode::updateNode()