+++ src/GUI/SetupColourConfigUI.hpp
@@ -6,12 +6,14 @@
 #include "../Client/ModuleNode.hpp"
 #include "../Client/CategoryNode.hpp"
 #include "../Utils/ColourConfig.hpp"
+#include "BetterSlider.hpp"
 
 using namespace geode::prelude;
 
 class SetupColourConfigUI : public geode::Popup<>, public ColorPickerDelegate
 {
     protected:
+        bool allowEffects = false;
         BackgroundSprite* bg = nullptr;
         ColourConfig defaultConfig;
         ColourConfig startConfig;
@@ -21,14 +23,31 @@ class SetupColourConfigUI : public geode::Popup<>, public ColorPickerDelegate
         CCLayerColor* endColour = nullptr;
         std::string previewChannel = "";
         CCControlColourPicker* picker = nullptr;
+        CCMenu* typeMenu = nullptr;
         std::map<CCMenuItemToggler*, ColourConfigType> configTypes = {};
+        CCNode* gradientPreviewContainer = nullptr;
+        std::unordered_map<CCLayerColor*, float> gradientPreviewSprites = {};
+        CCNode* gradientTimePreview = nullptr;
+        std::vector<CCNode*> gradientLines = {};
+        int selectedGradientLine = 0;
+        CCMenu* gradientLineConfigNode = nullptr;
+        CCSprite* gradientLineColour = nullptr;
+        BetterSlider* gradientLineLocation = nullptr;
+        CCMenuItemSpriteExtra* gradientAddStepBtn = nullptr;
+        CCMenuItemSpriteExtra* gradientDelStepBtn = nullptr;
 
     public:
-        static SetupColourConfigUI* create(std::function<void(ColourConfig)> onFinishFunc);
+        static SetupColourConfigUI* create(std::function<void(ColourConfig)> onFinishFunc, bool allowEffects = true);
 
         void onClose(CCObject* sender);
         void onChangeType(CCObject* sender);
         void onSetDefault(CCObject* sender);
+        void onUndoChanged(CCObject* sender);
+        void onSelectGradientLine(CCObject* sender);
+        void onChangeGradientLineColour(CCObject* sender);
+        void onGradientLocationSlider(CCObject* sender);
+        void onAddGradientStep(CCObject* sender);
+        void onDeleteGradientStep(CCObject* sender);
 
         void setStartConfig(ColourConfig config);
         void setDefaultConfig(ColourConfig config);
@@ -38,6 +57,11 @@ class SetupColourConfigUI : public geode::Popup<>, public ColorPickerDelegate
 
         void addTypeButtons(CCMenu* menu);
         void updateTypeButtons(CCMenuItemToggler* excluding);
+
+        void createGradientPreview();
+        void updateGradientPreview();
+        CCNode* createGradientLine(bool timePreview, ccColor3B col = ccWHITE, bool selected = false, int tag = 0);
+        void updateGradientLines();
 
         virtual void colorValueChanged(ccColor3B colour);
         virtual void update(float dt);